removed instances of Arc<Mutex<T>>, but now have to deal with model loading issue smh...
