+++ eucalyptus-core/src/lib.rs
@@ -5,3 +5,4 @@ pub mod logging;
 pub mod scripting;
 pub mod states;
 pub mod utils;
+pub mod spawn;
