--- src/dropbear.gleam
@@ -1,3 +1,0 @@
-/// Queries the attached entity. Returns a serialized string to be decoded. 
-@external(javascript, "dropbear", "queryAttachedEntity")
-fn raw_query_attached_entity() -> String
