+++ src/dropbear/entity.gleam
@@ -1,0 +1,5 @@
+import dropbear/dropbear
+
+pub fn query_current_entity() -> dropbear.QueryId {
+
+}
