+++ src/example.gleam
@@ -1,0 +1,9 @@
+import dropbear
+
+pub fn load() {
+  
+}
+
+pub fn update(dt: Int) {
+
+}
